requirements.txt
asyncpg==0.30.0
playwright==1.49.1
aiohttp==3.11.11
python-dotenv==1.0.1
apscheduler==3.11.0
jinja2==3.1.5
